Python script for IMEM initialization + fix issues
Usage: python3 mips_to_machine_code.py <path to asm file> <path to IMEM.mem>

There is a mips assembly file in test-utils that loads a 32-bit word, adds 7 to it, and stores it in another memory location. This test shows that the data path of this crypto-multiprocessor is fully functioning and correct. It's worth mentioning that the control path for this processor can still be improved.